New entrypoint for lnd

+++ lnd/v0.12.0-beta/entrypoint.sh
@@ -33,20 +33,18 @@ else
   echo "/lnd user and group is good !"
 fi
 
-if [ ! -z "${BITCOIND_RPCPASS}" ] && [ -f $LND_CONF_PATH ]; then
+if [ ! -z "${BITCOIND_RPCPASS}" ] && [ -f $LND_CONF_PATH ] && [ grep -q BITCOIND_RPCPASS $LND_CONF_PATH]; then
   sed "s/BITCOIND_RPCPASS/$BITCOIND_RPCPASS/g" $LND_CONF_PATH > tmp.conf
   cat tmp.conf > $LND_CONF_PATH
   rm tmp.conf
 fi
+if [ ! -z "${TOR_PASSWORD}" ] && [ -f $LND_CONF_PATH ] && [ grep -q TOR_PASSWORD $LND_CONF_PATH]; then
+  sed "s/BITCOIND_RPCPASS/$TOR_PASSWORD/g" $LND_CONF_PATH > tmp.conf
+  cat tmp.conf > $LND_CONF_PATH
+  rm tmp.conf
+fi
 
-if [ "$1" = "lnd" ]; then
-  if [ ! -z "${BITCOIND_RPCPASS}" ] && [ -f $LND_CONF_PATH ]; then
-    sed "s/BITCOIND_RPCPASS/$BITCOIND_RPCPASS/g" $LND_CONF_PATH > tmp.conf
-    cat tmp.conf > $LND_CONF_PATH
-    rm tmp.conf
-  fi
-  exec su-exec lnd $@
-elif [ "$1" = "lnd-cli" ] || [ "$1" = "lnd-tx" ]; then
+if [ "$1" = "lnd" ] || [ "$1" = "lncli" ]; then
   exec su-exec lnd "$@"
 else
   exec "$@"
